Remove Exceptions, move Property into Primitive.kt as inner class
