+++ core/src/main/kotlin/net/bladehunt/window/core/layer/ArrayLayerImpl.kt
@@ -23,7 +23,6 @@
 
 package net.bladehunt.window.core.layer
 
-import net.bladehunt.window.core.exception.WindowException
 import net.bladehunt.window.core.ext.forEach
 import net.bladehunt.window.core.ext.forEachNotNull
 import net.bladehunt.window.core.util.IntPair
@@ -62,7 +61,6 @@ class ArrayLayerImpl<Pixel>(
     }
 
     override fun copyTo(other: Layer<Pixel>) {
-        if (other.size != size) throw WindowException("Layers must be the same size to copy")
         pixels.forEach { x, y, value ->
             if (value == null) other.remove(x, y)
             else other[x, y] = value
